refactor/ Component name synchronized and unused files deleted
