+++ client/src/components/cards/ProductCardHorizontal.js
@@ -1,67 +1,104 @@
-import moment from "moment";
-import { useCart } from "../../context/cart";
+// 👻 Developed by DanBi Choi on July 26th, 2023.
+// 👻 Developed by DanBi Choi on July 29th, 2023. (Add To Cart Button)
+// 👻 Developed by DanBi Choi on Aug 20th, 2023. (Add To Cart Modulized)
+// -----------------------------------------------------
+import { useNavigate } from "react-router-dom";
+import "../../styles/components/cards/ProductCardHorizontal.scss";
+import { AiFillStar } from "react-icons/ai";
+import useWindowWidth from "../../hooks/useWindowWidth";
+import useAddToCart from "../../hooks/useAddToCart";
 
-export default function ProductCardHorizontal({ p, remove = true }) {
-    // context
-    const [cart, setCart] = useCart();
+export default function ProductCardHorizontal({ product, modify = false }) {
+    // hooks
+    const windowWidth = useWindowWidth();
+    const navigate = useNavigate();
+    const addtoCart = useAddToCart();
 
-    const removeFromCart = (productId) => {
-        let myCart = [...cart];
-        let index = myCart.findIndex((item) => item._id === productId);
-        myCart.splice(index, 1);
-        setCart(myCart);
-        localStorage.setItem("cart", JSON.stringify(myCart));
+    const handleAddToCart = (e) => {
+        e.preventDefault();
+        addtoCart(product);
+    };
+
+    const handleLinkClick = (e) => {
+        e.preventDefault();
+        navigate(`/shop/product/${product?.slug}`);
     };
 
     return (
-        <div
-            className="card mb-3"
-            // style={{ maxWidth: 540 }}
-        >
-            <div className="row g-0">
-                <div className="col-md-4">
-                    <img
-                        src={`${process.env.REACT_APP_API}/product/images/${p._id}`}
-                        alt={p.name}
-                        style={{
-                            height: "150px",
-                            width: "150px",
-                            objectFit: "cover",
-                            marginLeft: "-12px",
-                            borderRopRightRadius: "0px",
-                        }}
-                    />
-                </div>
-                <div className="col-md-8">
-                    <div className="card-body">
-                        <h5 className="card-title">
-                            {p.name}{" "}
-                            {p?.price?.toLocaleString("en-US", {
-                                style: "currency",
-                                currency: "USD",
-                            })}
-                        </h5>
-                        <p className="card-text">{`${p?.description?.substring(
-                            0,
-                            50
-                        )}..`}</p>
-                    </div>
-                </div>
+        <div className="card-container">
+            <div className="img" onClick={handleLinkClick}>
+                <img
+                    src={`${process.env.REACT_APP_API}/product/images/${product._id}`}
+                    alt={product.title}
+                />
+            </div>
+            <div className="text d-flex flex-column justify-content-between align-items-start">
+                <h2>{product?.category?.name}</h2>
+                <h3 onClick={handleLinkClick}>
+                    {windowWidth > 450 &&
+                        (product?.title?.length > 35
+                            ? product?.title?.substring(0, 35) + "..."
+                            : product?.title)}
+                    {windowWidth < 450 &&
+                        (product?.title?.length > 25
+                            ? product?.title?.substring(0, 25) + "..."
+                            : product?.title)}
+                </h3>
+                <h5 onClick={handleLinkClick}>
+                    {windowWidth > 1200 &&
+                        (product?.description?.length > 140
+                            ? product?.description.substring(0, 140) + "..."
+                            : product?.description)}
+                    {windowWidth < 1200 &&
+                        (product?.description?.length > 70
+                            ? product?.description.substring(0, 70) + "..."
+                            : product?.description)}
+                </h5>
 
-                <div className="d-flex justify-content-between">
-                    <p className="card-text">
-                        <small className="text-muted">
-                            Listed {moment(p.createdAt).fromNow()}
-                        </small>
-                    </p>
-                    {remove && (
-                        <p
-                            className="text-danger mb-2 pointer"
-                            onClick={() => removeFromCart(p._id)}
-                        >
-                            Remove
-                        </p>
-                    )}
+                <div
+                    className="text-bottom d-flex flex-row justify-content-between align-items-center"
+                    style={{ width: "100%" }}
+                >
+                    <h4>${product?.price}</h4>
+                    <div className="d-flex flex-row justify-content-between">
+                        <div className="product-rate-box d-flex flex-row justify-content-between align-items-center">
+                            <h6>
+                                {product?.reviewRate
+                                    ? product?.reviewRate.toFixed(2)
+                                    : "No Rating"}
+                            </h6>
+                            <h6>
+                                <AiFillStar style={{ fill: "#ffbf35" }} />
+                            </h6>
+                            <h6>
+                                (
+                                {product?.reviewNumber
+                                    ? product?.reviewNumber
+                                    : "0"}
+                                )
+                            </h6>
+                        </div>
+                        {modify ? (
+                            <button
+                                className="btn btn-primary"
+                                onClick={(e) => {
+                                    e.preventDefault();
+                                    navigate(
+                                        `/dashboard/admin/product/update/${product.slug}`
+                                    );
+                                }}
+                            >
+                                Modify Product
+                            </button>
+                        ) : (
+                            <button
+                                className="btn btn-primary"
+                                onClick={handleAddToCart}
+                            >
+                                Add to Cart
+                            </button>
+                        )}
+                    </div>
                 </div>
             </div>
         </div>
