+++ client/src/styles/components/cards/ProductCardHorizontal.scss
@@ -79,6 +79,7 @@
             color: $color-gray;
             font-size: $desktop-font-medium;
             margin-bottom: 20px;
+            cursor: pointer;
         }
 
         h4 {
